+++ tests/fakes/__init__.py
@@ -11,7 +11,12 @@ from tests.fakes.audit import FakeAuditLogger
 from tests.fakes.balance import FakeBalanceConfig
 from tests.fakes.clan_repo import FakeClanMembershipRepository, FakeClanRepository
 from tests.fakes.clock import FakeClock
-from tests.fakes.dau import FakeDauCounter, FakeDauLimit
+from tests.fakes.dau import (
+    DauAlertEvent,
+    FakeDauCounter,
+    FakeDauLimit,
+    FakeDauThresholdAlerter,
+)
 from tests.fakes.idempotency import FakeIdempotencyKey
 from tests.fakes.player_repo import FakePlayerRepository
 from tests.fakes.random import FakeRandom
@@ -19,6 +24,7 @@ from tests.fakes.signup_queue import FakeSignupQueueRepository
 from tests.fakes.uow import FakeUnitOfWork
 
 __all__ = [
+    "DauAlertEvent",
     "FakeAdminRepository",
     "FakeAuditLogger",
     "FakeBalanceConfig",
@@ -27,6 +33,7 @@ __all__ = [
     "FakeClock",
     "FakeDauCounter",
     "FakeDauLimit",
+    "FakeDauThresholdAlerter",
     "FakeIdempotencyKey",
     "FakePlayerRepository",
     "FakeRandom",